Workaround for VerifyError on API 17...

+++ ui-preview/src/main/java/l/files/ui/preview/MediaTypeCache.java
@@ -9,7 +9,7 @@ import l.files.fs.Path;
 final class MediaTypeCache extends PersistenceCache<String> {
 
     MediaTypeCache(Path cacheDir) {
-        super(cacheDir, (byte) 1);
+        super(cacheDir, 1);
     }
 
     @Override
